+++ gasoline/src/java/com/stepup/gasoline/qt/bean/RouteBean.java
@@ -12,7 +12,7 @@ public class RouteBean {
 
     private int id;
     private String name;
-    private float distance;
+    private int distance;
     private String note;
 
     public void setId(int id) {
@@ -31,11 +31,11 @@ public class RouteBean {
         this.name = name;
     }
 
-    public float getDistance() {
+    public int getDistance() {
         return distance;
     }
 
-    public void setDistance(float distance) {
+    public void setDistance(int distance) {
         this.distance = distance;
     }
