+++ gasoline/src/java/com/stepup/gasoline/qt/fixedasset/FixedAssetFormBean.java
@@ -20,8 +20,8 @@ public class FixedAssetFormBean extends org.apache.struts.action.ActionForm {
     private String unitName;
     private String startDate;
     private int quantity;
-    private float rate;
-    private float monthCount;
+    private int rate;
+    private int monthCount;
     private int status;
     private String statusName;
     private double price;
@@ -106,19 +106,19 @@ public class FixedAssetFormBean extends org.apache.struts.action.ActionForm {
         this.quantity = quantity;
     }
 
-    public float getRate() {
+    public int getRate() {
         return rate;
     }
 
-    public void setRate(float rate) {
+    public void setRate(int rate) {
         this.rate = rate;
     }
 
-    public float getMonthCount() {
+    public int getMonthCount() {
         return monthCount;
     }
 
-    public void setMonthCount(float monthCount) {
+    public void setMonthCount(int monthCount) {
         this.monthCount = monthCount;
     }
