+++ gasoline/src/java/com/stepup/gasoline/qt/gasimport/AddGasImportAction.java
@@ -88,9 +88,9 @@ public class AddGasImportAction extends SpineAction {
                 if (id == 0) {
                     GasImportDetailBean bean = new GasImportDetailBean();
                     bean.setShellId(NumberUtil.parseInt(formBean.getShellId()[i], 0));
-                    bean.setQuantity(NumberUtil.parseFloat(formBean.getQuantity()[i], 0));
-                    bean.setPrice(NumberUtil.parseFloat(formBean.getPrice()[i], 0));
-                    bean.setAmount(NumberUtil.parseFloat(formBean.getAmount()[i], 0));
+                    bean.setQuantity(NumberUtil.parseInt(formBean.getQuantity()[i], 0));
+                    bean.setPrice(NumberUtil.parseInt(formBean.getPrice()[i], 0));
+                    bean.setAmount(NumberUtil.parseInt(formBean.getAmount()[i], 0));
                     bean.setGasImportId(formBean.getId());
                     gasDAO.insertGasImportDetail(bean);
                 } else {
@@ -107,15 +107,15 @@ public class AddGasImportAction extends SpineAction {
                         arrDetail.remove(j);
                         if (oldBean.getQuantity() != NumberUtil.parseDouble(formBean.getQuantity()[i], 0)) {
                             isUpdate = true;
-                            oldBean.setQuantity(NumberUtil.parseFloat(formBean.getQuantity()[i], 0));
+                            oldBean.setQuantity(NumberUtil.parseInt(formBean.getQuantity()[i], 0));
                         }
                         if (oldBean.getPrice()!= NumberUtil.parseDouble(formBean.getPrice()[i], 0)) {
                             isUpdate = true;
-                            oldBean.setPrice(NumberUtil.parseFloat(formBean.getPrice()[i], 0));
+                            oldBean.setPrice(NumberUtil.parseInt(formBean.getPrice()[i], 0));
                         }
                         if (oldBean.getAmount()!= NumberUtil.parseDouble(formBean.getAmount()[i], 0)) {
                             isUpdate = true;
-                            oldBean.setAmount(NumberUtil.parseFloat(formBean.getAmount()[i], 0));
+                            oldBean.setAmount(NumberUtil.parseInt(formBean.getAmount()[i], 0));
                         }
                         if (isUpdate) {
                             gasDAO.updateGasImportDetail(oldBean);
